requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-multipart==0.0.6
pdf2image==1.16.3
PyPDF2==3.0.1
reportlab==3.6.12
Pillow==9.5.0
loguru==0.7.0
pydantic==1.10.7
pyyaml==6.0
googletrans==4.0.0-rc1 